test: Implemented test export census
Implemented a new class test called CensusExportImport in tests.py that contains the test case "test_export_census" which tries to export a census first with a user not admin and then with a user that is part of the staff.
The necessary imports have been made and the necessary set up and tear conditions have been implemented.